requirements.txt
pytest==7.1.2
allure-pytest==2.9.45
selene==1.0.2
python-dotenv==0.20.0